+++ src/client/adapters/rdfjs/n3/mod.ts
@@ -1,1 +1,0 @@
-export * from "./proxy-store.ts";