+++ src/client/quad-store/mod.ts
@@ -1,5 +1,6 @@
 export * from "./quad-store-interface.ts";
 export * from "./patch.ts";
+export * from "./merge-patches.ts";
 export * from "./quad-filter.ts";
 export * from "./is-textual-literal.ts";
 export * from "./hash-quad.ts";