+++ src/client/quad-store/n3/mod.ts
@@ -1,0 +1,1 @@
+export * from "./create-proxied-n3-store.ts";